don't pollute ActiveRecord::Base until validations are enabled
Also add schema validations auto-create on demand.
With that approach we don't unnecessary alias_method_chain when
schema_validations are disabled.

It causes some testing problems. Enabling a feature means extending
a class so there is no way to unextend and rewind the changes on the class.

At the moment specs for schema_validations pass when runned standalone but fail when
running around other specs.