bug fix: per-table override of :on_update and :on_delete was being ignored.  added spec tests for the various setting & overriding of :on_update and :on_delete.
(i was actually in the process of getting rid of the global set_default_update_and_delete_actions! method anyway as a small step towards issue #22 'stronger modularization', and noticed that it wasn't using the per-table overrides and that the specs weren't testing that case.  so added the specs, saw that they failed, then got rid of the global method.)