mysql adapter conformity/bug fix: tables_with_active_schema now takes name and other args to pass through to tables_without_active_schema
this fixes a bunch of argument errors that started cropping up. whether
it's one or two args -- just (name) or (name, database) -- depends on the
version.

in practice those args are seem to be  nil, which is the default value. i
suspect that in previous versions of active_record they weren't being
passed but now they are.

also passing the name arg to views() and from views() through to execute, which is what it's for.  though again, i think it's generally nil in practice.