Allow initial space and \n chars in PG views
Three of the tests were failing for me.  All of them parse output from
the pg_dump command, and my version of Postgres was putting a space
before the word SELECT in the view, as well as embedded \n characters
inside of them.  I changed the regexps to permit those new
peculiarities, while still allowing everything that was previously OK.